Remove polluting scope
